feature(network): drop MacVLAN networking
Due to the limitations of the MacVLAN networks (containers can't
communicate with the host, host can't communicate with the
containers and so on) I decided to drop the support for MacVLAN in all
containers in favor of direct port mappings whenever is necessary.

Most of the services are already being exposed through the traefik
reverse proxy and should not have any impact.

Signed-off-by: Filipe Utzig <filipe@gringolito.com>